adding image reading - finished library
